Making the confidential client object global  (#14)
* Making the confidential client object global 

Apparently, whoever is referring the sample is creating the confidential client as a local variable. That is not desired. It can be created as a global variable and then later call the acquiretoken method to use of the cache.

* Making the confidential client object global 

Apparently, whoever is referring the sample is creating the confidential client as a local variable. That is not desired. It can be created as a global variable and then later call the acquiretoken method to use of the cache.